Очень большое количество фикса

+++ libft/src/ft_fill_zeroes.c
@@ -6,7 +6,7 @@
 /*   By: idunaver <idunaver@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/21 10:50:35 by ramory-l          #+#    #+#             */
-/*   Updated: 2019/06/15 17:09:31 by idunaver         ###   ########.fr       */
+/*   Updated: 2019/09/16 15:50:50 by idunaver         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 static int	for_mode(char *buffer, char mode, size_t size)
 {
-	int		i;
+	size_t		i;
 
 	i = 0;
 	if (mode == '>')
@@ -27,7 +27,7 @@ static int	for_mode(char *buffer, char mode, size_t size)
 			i++;
 		}
 	}
-	return (i);
+	return ((int)i);
 }
 
 static int	back_mode(char *buffer, char mode, size_t size, int len)
