+++ libft/src/ft_printf/src/support/ft_fill_bf.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_fill_bf.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ramory-l <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: idunaver <idunaver@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/02 15:24:51 by ramory-l          #+#    #+#             */
-/*   Updated: 2019/06/15 18:17:15 by ramory-l         ###   ########.fr       */
+/*   Updated: 2019/09/16 15:54:35 by idunaver         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	ft_fill_bf(t_printf *data, t_buffer *bf)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	i = 0;
 	j = data->len + bf->s;
