+++ libft/src/ft_printf/src/support/ft_fill_bf_zeroes.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_fill_bf_zeroes.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ramory-l <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: idunaver <idunaver@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/02 15:15:29 by ramory-l          #+#    #+#             */
-/*   Updated: 2019/06/15 13:59:32 by ramory-l         ###   ########.fr       */
+/*   Updated: 2019/09/16 15:56:39 by idunaver         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ void	ft_fill_zeroes_di(t_printf *data, t_buffer *bf)
 
 void	ft_fill_zeroes_oux(t_printf *data, t_buffer *bf)
 {
-	int zeroes;
+	long long zeroes;
 
 	zeroes = data->width - data->len + bf->s;
 	if (data->flags & FLAG_OCT)
